Fix infinite, no-progress loop on initial empty database
If the database is empty, we use the current time as the _after_ time
for recently played. However, if the database is empty and we ask for
things after the current time, Spotify will always give us empty (since
the currently playing song will not appear). When we come around to the
loop again, we'll use a _new_ recent time, which will disqualify any
songs that played in between last time.

The effect is we end up never getting anything back from Spotify.

Instead, we now use the initialization time, which should be around the
time the app starts. This way, we'll make sure to eventually get
something back.